Add HTTP 451 amendments to Content API approach
Includes some copy introducing a revised approach to Gone types
considered from the context of the 451 HTTP status code.

This has been done with minimal edits to the document since the deadline
for the RFC is very soon and the audience for this RFC is likely already
acquainted with the main copy.