Change from forcing lowercase to single case
Based on feedback and discussions, this commit amends the RFC to not force all base paths to be lowercase, but instead only allow one case to be registered.

This will allow for situations where a non lowercase base path is preferred as canonical but will still prevent two cases of the same base path having different meanings.

It also allows disk-based files (such as assets) to have their routes registered with the same case as the filename to avoid having an exception for files.