Add guidance on resolving errors
From @MatMoore:

> Another thing that possibly ties in with this RFC, is our workflow
for sentry. It's possible to resolve and silence errors in sentry and
it's also possible to customise the conditions when an event triggers a
notification. By default, we're only notifying on the first occurrence
of an error, and after that the only way it gets fixed is if people
look at sentry and see it.