Update RFC to reflect revised featured_attachments approach
This amends the RFC to resolve some misunderstandings with the approach
we'd be taking for this work. This change reflects a desire to no longer
pass HTML or Govspeak into content items for attachments and instead
pass in an id to reference which attachment should be rendered, the
rendering applications will then take this id and use a component to
render the resulting attachment.