Water down the difficulty of upstart/systemd
In theory puppet supports both upstart and systemd, and it should
magically handle the switch for us. There are a few references to
upstart specifically in govuk-puppet, but not so many that it looks like
it would be a massive job.

In practice I think it would be a bit harder than that, but without
trying it I can't know for sure. I'm watering down the wording so we
can still call this out as a thing without making it sound like it's a
main piece of the problem.