Clarify "containers" not necessarily images
The intention here is not to rule out any tool that lets us build
containers. I'm stopping short of calling out things like Guix as
options, because it's a rabbit hole I'd prefer not to go down in this
RFC. I think we should give reasonable consideration to all the tools
available to us when we get to the point where we can make this
decision.