Add additional constraint: auto-merging to happen in office hours
This is in response to RFC feedback. There is a consensus that
we should only automate merging PRs to happen during office hours,
to allow time for developers to step in and revert changes if it
turns out our other safety measures don't prevent a bad release
from getting through.