Refine constraint: auto-merging to happen in office hours
This commit tightens up the language around the days when auto
merging should happen (i.e. 9:30am, Monday-Friday). The 9am-4pm
window was confusing and gave developers unnecessary freedom to
e.g. configure their apps' Dependabot schedule to run at 4pm when
there is no clear advantage to doing that over 9:30am.

It also removes the suggestion of a manually enabled/disabled
feature flag to get around the bank holiday issue, and instead
proposes using GOV.UK's Bank Holidays API to validate that the
current date is not a bank holiday. This will add some custom
scripting to the resulting GitHub Action, but if we package the
action up as proposed in the RFC, it should not mean any additional
complexity in the configs rolled out across GOV.UK.

See feedback that led to this change:
https://github.com/alphagov/govuk-rfcs/pull/156#pullrequestreview-1299665416