Add note clarifying the effect of schedule time on security PRs
Just to reassure anyone who may mistakenly believe that setting
Dependabot to run at 9:30am every day might delay Dependabot in
raising security-related PRs for external dependencies: it won't.

As part of this clarification, I did go down a slight rabbit hole
around what Dependabot would do for a security-related _internal_
dependency PR, but it's so niche and unlikely to materialise that
I think it's not even helpful to include in the RFC body. I'm
happy to include it in this commit message though.

In theory, an internal dependency update that is somehow flagged
as a security fix could end up being auto-merged outside of office
hours. However, [according to GitHub][about-dependabot-security-updates]:

- Security PRs are raised when a vulnerable dependency from the
  [GitHub Advisory Database][] is detected in the dependency graph,
  and where there is a newer dependency version that fixes it.
- As far as we know, a GOV.UK-maintained repository has never
  ended up in the database, so this is an extremely unlikely edge
  case. And even if it were to happen, the risk of leaving our
  applications unpatched would probably outweigh the risk of
  auto-merging the patch out of hours.
- I can't even envisage a scenario where there'd be no human
  involvement whatsoever. Say rubocop-govuk v30.0.0 has a security
  vulnerability and ends up being included in the GitHub Advisory
  Database. It's fixed in v30.0.1. Dependabot raises a bunch
  of 30.0.1 PRs out of hours and they all get auto-merged because
  they're internal dependency PRs.
  Someone had to have made 30.0.1 to fix the security issue, and
  if they've done that, they're presumably going to stick around
  and try to apply the fix everywhere.
- This is on top of the already pretty good safety nets of 95%
  code coverage and a passing CI being pre-requisites for auto
  merge.

Basically this was too big of a distraction to include in the RFC,
but I do want to show I've considered it - hence this commit!

[about-dependabot-security-updates]: https://docs.github.com/en/code-security/dependabot/dependabot-security-updates/about-dependabot-security-updates
[GitHub Advisory Database]: https://github.com/advisories