Propose using bespoke auto-merge service instead of GitHub Actions
Following [feedback highlighting the risk of privilege escalation
with the use of GitHub Actions][discussion-on-implementation],
we've decided to decouple the auto merging mechanism from GitHub.

This minimises the attack vector: it would no longer be possible
for a compromised GitHub account to sidestep the mandatory review
process through leveraging access tokens or the GitHub Actions bot
permissions, as the former wouldn't be held in GitHub and the latter
would not be enabled.