Technical Direction in Publishing 2023
This RFC is to document opinions and approaches that have been discussed
in the Publishing directorate over the last year or so. For those
who've been in publishing a while, there isn't expected to be any huge
surprises (and please feedback if there is).

The motivation for putting this together was that: Publishing has grown
rapidly and continues to have frequent people changes; Kevin and Al
(lead tech in Publishing) were about to go months without working
together and wanted to document the current thinking; and that we have outdated
directional information in [RFC-95][1].

It's welcome for other directorates than Publishing to have opinions on
these principles, or embrace them, however we've only sought product
agreement within Publishing.

[1]: https://github.com/alphagov/govuk-rfcs/blob/main/rfc-095-long-term-future-apps.md