Add out of date and superseding note to RFC 95
With the introduction of RFC 161 much of the information in this RFC is
now superseded. I thought it was worth adding the extra qualifier that
the information in this no longer represents current direction of travel
given not all of this is superseded but much of it would be considered
outdated.