Add RFC metadata as YAML frontmatter
We reviewed the latest 42 RFCs with senior tech to check on their
statuses (i.e. accepted / superseded etc.) and implementation status
(i.e. done / in progress etc.).

This adds metadata to each RFC as YAML frontmatter.

Also add a script to the README to show how this can be used in a
programmatic way. The results of running this script can be seen in this
spreadsheet:

https://docs.google.com/spreadsheets/d/1bYF7r4IC7owbE7NrbFMZYJiuqsPZeQka1fzYD4apx9U/edit#gid=166323961